:recycle: Extract enforce not blank into assertions.py. Also, remove src directories
