+++ source/YugiAcademy/YuGiOh/cards/src/spell_card.py
@@ -1,13 +1,7 @@
 from django.db import models
 
 from YuGiOh.models import Card
-
-from utils.src.instance_creation_failed import InstanceCreationFailed
-
-
-def enforce_not_blank(string, name):
-    if not string.strip():
-        raise InstanceCreationFailed(f'{name} must not be blank.')
+from assertions.assertions import enforce_not_blank
 
 
 class SpellCard(Card):
