+++ source/YugiAcademy/assertions/instance_creation_failed.py
